# crates/atuin-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atuin Sync Server API",
    "description": "HTTP API for the atuin sync server (crates/atuin-server). Schemas are derived from the Rust types in atuin-common, so they mirror their serde output exactly. All error responses use the ErrorResponse shape {\"reason\": string}. Every response carries an `Atuin-Version` header with the server's cargo version; the client uses it to refuse syncing across major versions. If the server is configured with a URL path prefix, it must be included in the client base URL.",
    "contact": {
      "name": "Ellie Huxtable",
      "email": "ellie@atuin.sh"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/account": {
      "delete": {
        "tags": [
          "handlers::user"
        ],
        "operationId": "delete_account",
        "responses": {
          "200": {
            "description": "Account deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/account/password": {
      "patch": {
        "tags": [
          "handlers::user"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Not authenticated, or current password incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "handlers::user"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logged in; returns a session token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "handlers::user"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Registered; returns a session token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Registration rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "handlers::user"
        ],
        "operationId": "get_user",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username to look up",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "User not found or request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        }
      },
      "ChangePasswordResponse": {
        "type": "object"
      },
      "DeleteUserResponse": {
        "type": "object"
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "session"
        ],
        "properties": {
          "auth": {
            "type": "string",
            "description": "Auth type: \"hub\" for Hub API tokens, \"cli\" for legacy CLI session tokens.\nOld servers that don't return this field will deserialize as None.",
            "nullable": true
          },
          "session": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "username",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "required": [
          "session"
        ],
        "properties": {
          "auth": {
            "type": "string",
            "description": "Auth type: \"hub\" for Hub API tokens, \"cli\" for legacy CLI session tokens.\nOld servers that don't return this field will deserialize as None.",
            "nullable": true
          },
          "session": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Session token presented as `Authorization: Token <session-token>`. This server accepts only the `Token` scheme (exactly one space separator); any other scheme is rejected with 400 {\"reason\":\"invalid authorization header encoding\"}. The full header value including the `Token ` prefix must be supplied."
      }
    }
  }
}